srcs/services/django/conf/requirements.txt
Django>=4.2,<5.0
gunicorn>=20.1,<21.0
psycopg2-binary>=2.9,<3.0
channels==3.0.5
channels-redis
whitenoise
django-environ
djangorestframework
djangorestframework-simplejwt
django-prometheus==2.3.1
django-cors-headers
requests
django-allauth